Extract SearchInfo interface and SearchScope enum
