+++ test/components/structures/RoomView-test.tsx
@@ -70,7 +70,7 @@ import WidgetUtils from "../../../src/utils/WidgetUtils";
 import { WidgetType } from "../../../src/widgets/WidgetType";
 import WidgetStore from "../../../src/stores/WidgetStore";
 import { ViewRoomErrorPayload } from "../../../src/dispatcher/payloads/ViewRoomErrorPayload";
-import { SearchScope } from "../../../src/components/views/rooms/SearchBar";
+import { SearchScope } from "../../../src/Searching";
 
 const RoomView = wrapInMatrixClientContext(_RoomView);
